requirements.txt
kafka-python==2.2.15
python-dotenv==1.1.1
rel==0.4.9.20
websocket-client==1.8.0
